Module Src Code

+++ Hello_World/Hello_mod.c
@@ -1,5 +1,6 @@
 #include <linux/init.h>
 #include <linux/module.h> 
+#include <linux/types.h>
 #include <linux/moduleparam.h>
 #include <linux/kernel.h>
 #include <linux/printk.h> 
@@ -8,14 +9,14 @@ MODULE_LICENSE("GPL");
 MODULE_AUTHOR("AKhadhraoui47");
 MODULE_DESCRIPTION("Hello World Module");
 
-static unsigned int argc_k = 0;
-static short int argv_k[2] = {-1, -1};
+static uint32_t argc_k = 0;
+static uint16_t argv_k[2] = {-1, -1};
 module_param_array(argv_k, short, &argc_k, S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP );
 MODULE_PARM_DESC(argv_k, "My Array of short Integers");
 
 static int __init hello_init(void)
 {
-    short int i = 0;
+    uint8_t i = 0;
     pr_info("Hello, world\n");
     for (i = 0; i < ARRAY_SIZE(argv_k); i++)
     {
